Update: Se refactorizo JDBCController como un servicio
