+++ src/main/java/module-info.java
@@ -7,4 +7,5 @@ module com.asesinosdesoftware.javeparking {
     opens com.asesinosdesoftware.javeparking to javafx.fxml;
     opens com.asesinosdesoftware.javeparking.controller to javafx.fxml;
     exports com.asesinosdesoftware.javeparking;
+    opens com.asesinosdesoftware.javeparking.services to javafx.fxml;
 }
